+++ tools/sector_scraper.js
@@ -1,5 +1,14 @@
 import Hackmud from "../src/hackmud";
 
+/**
+ * Scrapes sectors looking for corp scripts.
+ * @param {Object} context Context.
+ * @param {Object} args Arguments passed to the script.
+ * @param {Number} [args.tier=1] What tier of corps to look for.
+ * @param {Number} [args.start=0] Where in the array of sectors to start looking.
+ *
+ * @returns {Array} Array of corp scripts
+ */
 export default function (context, args) {
     let
         sectorStart = args.start || 0,
@@ -8,20 +17,24 @@ export default function (context, args) {
         sector = "s",
         results = [],
         sectorList = [],
+        runFullSec = params => Hackmud.fs.scripts.fullsec(params),
+        runHighSec = params => Hackmud.fs.scripts.highsec(params),
+        runMidSec = params => Hackmud.fs.scripts.midsec(params),
         i;
+
 
     const runSectors = sectorFunction => {
         i = 0;
         sectorList = sectorFunction().slice(sectorStart, sectorStart + 11);
 
-        for (;sector; i++) {
+        for (;sectorList[i]; i++) {
             sector = sectorList[i];
 
             Hackmud.ms.chats.join({ channel: sector });
 
             scriptList = scriptList.concat(sectorFunction({ sector: sector })
                 .filter(scriptName =>
-                    /\.public|private|member|access/.exec(scriptName)
+                    /\.public|private|member/.exec(scriptName)
                 ));
 
             Hackmud.ms.chats.leave({ channel: sector });
@@ -32,14 +45,14 @@ export default function (context, args) {
 
     if (tier == 1) {
         results = results.concat(
-            runSectors(params => Hackmud.fs.scripts.fullsec(params))
+            runSectors(runFullSec)
         );
     }
 
     if (tier == 2) {
         results = results.concat(
-            runSectors(params => Hackmud.fs.scripts.highsec(params)),
-            runSectors(params => Hackmud.fs.scripts.midsec(params))
+            runSectors(runHighSec),
+            runSectors(runMidSec)
         );
     }